payments and orders lists are finished

+++ src/main/kotlin/org/order/data/tables/Menus.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.dao.IntIdTable
 
 object Menus: IntIdTable() {
     val name = varchar("name", 255)
-    val cost = integer("cost")
+    val cost = float("cost")
     val active = bool("is_active")
     val schedule = integer("schedule")
 }
